Recreate migrations and clean apps
