--- bazaar/apps/customer/tests/__init__.py
@@ -1,1 +1,0 @@
-from shib import *
