+++ bazaar/apps/tests/__init__.py
@@ -1,0 +1,1 @@
+from apps.tests.shib import *
